Revise dto, entity, and schema.sql

+++ spacemission/src/main/java/tr/edu/bilkent/spacemission/dto/CompanyDto.java
@@ -15,7 +15,7 @@ public class CompanyDto {
     String companyName;
     String companyMail;
     String companyCountry;
-    long companyBudget;
+    double companyBudget;
     String companyType;
 
 }
